Checked temporary folder still exists before deleting it (in case 2 instances of Sweet Home 3D run at the same time)
git-svn-id: https://svn.code.sf.net/p/sweethome3d/code/trunk/SweetHome3D@2845 03e323e1-6c6c-46a5-9f71-c4ce5d8c6204

+++ src/com/eteks/sweethome3d/tools/OperatingSystem.java
@@ -186,8 +186,10 @@ public class OperatingSystem {
               public void run() {
                 long now = System.currentTimeMillis();
                 for (File siblingTemporaryFolder : siblingTemporaryFolders) {
-                  if (now - siblingTemporaryFolder.lastModified() > age) {
-                    for (File temporaryFile : siblingTemporaryFolder.listFiles()) {
+                  if (siblingTemporaryFolder.exists()
+                      && now - siblingTemporaryFolder.lastModified() > age) {
+                    File [] temporaryFiles = siblingTemporaryFolder.listFiles();
+                    for (File temporaryFile : temporaryFiles) {
                       temporaryFile.delete();
                     }
                     siblingTemporaryFolder.delete();
